requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
clearbit==0.1.7
click==7.1.1
Django==3.0.5
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-hstore==1.4.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
environs==7.3.1
idna==2.9
itsdangerous==1.1.0
MarkupSafe==1.1.1
marshmallow==3.5.1
passlib==1.7.2
psycopg2-binary==2.8.5
pycryptodome==3.9.7
pyhunter==1.7
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
Werkzeug==1.0.1
WTForms==2.2.1
